backend/requirements.txt
# Django Core
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ

# Database
psycopg2-binary

# API & Validation
Pillow

# Dev Tools
black
flake8

#uv pip install -r requirements.txt